fix: complete reel reorganization and fix namespace collision
- Copy sitecustomize.py to reel/ for coverage support
- Remove root .coveragerc and sitecustomize.py (now in reel/)
- Update root Makefile (simplified, tests now in reel/)
- Update root .gitignore (remove old .env patterns)
- Rename reel/grpc to reel/audio_grpc to avoid grpcio namespace collision
- Update all imports from reel.grpc to reel.audio_grpc
- Update Dockerfile COPY paths for renamed directory
- Add PYTHONPATH setup in test helpers for subprocess execution
- Fix reel/Makefile test path (tests/ not reel/tests/)
