--- .coveragerc
@@ -1,17 +1,0 @@
-[run]
-branch = True
-parallel = True
-source =
-    audio_to_text
-    audio_to_text_backend
-    audio_to_text_grpc
-    render_text_video
-    domain
-    service
-
-[report]
-show_missing = True
-skip_covered = True
-omit =
-    sitecustomize.py
-    audio_to_text_grpc/*_pb2*.py
