+++ .gitignore
@@ -21,9 +21,6 @@ coverage*.html
 # Managed by gix gitignore workflow
 .env
 .env.*
-!.env.audio_to_text_grpc.example
-!.env.audio_to_text_backend.example
-!.env.audio_to_text_ui.example
 .DS_Store
 qodana.yaml
 .idea/
