+++ reel/Makefile
@@ -6,7 +6,7 @@ COVERAGE_CONFIG ?= $(CURDIR)/.coveragerc
 test:
 	UV_CACHE_DIR="$(UV_CACHE_DIR)" uv run --no-project --with pytest \
 	    --with pillow --with grpcio --with grpcio-health-checking --with protobuf \
-	    -- python -m pytest -q reel/tests
+	    -- python -m pytest -q tests
 
 lint:
 	uv run --no-project --with mypy -- mypy --strict reel/domain reel/service
