+++ reel/docker/backend/Dockerfile.dev
@@ -2,8 +2,8 @@ FROM ghcr.io/astral-sh/uv:python3.12-bookworm AS builder
 
 WORKDIR /app
 COPY . /app
-RUN uv run --no-project /app/audio_to_text_backend.py --help >/dev/null \
-  && uv run --no-project /app/audio_to_text_grpc.py --help >/dev/null
+RUN uv run --no-project /app/reel/backend_server.py --help >/dev/null \
+  && uv run --no-project /app/reel/grpc_server.py --help >/dev/null
 
 FROM python:3.12-slim
 
@@ -17,11 +17,12 @@ WORKDIR /app
 
 COPY --from=builder /usr/local/bin/uv /usr/local/bin/uv
 COPY --from=builder /root/.cache/uv /root/.cache/uv
-COPY audio_to_text_backend.py /app/audio_to_text_backend.py
-COPY audio_to_text_backend /app/audio_to_text_backend
-COPY audio_to_text.py /app/audio_to_text.py
-COPY audio_to_text_grpc /app/audio_to_text_grpc
+COPY reel/__init__.py /app/reel/__init__.py
+COPY reel/backend_server.py /app/reel/backend_server.py
+COPY reel/backend /app/reel/backend
+COPY reel/audio_to_text.py /app/reel/audio_to_text.py
+COPY reel/audio_grpc /app/reel/audio_grpc
 
-RUN chmod +x /app/audio_to_text_backend.py
+RUN chmod +x /app/reel/backend_server.py
 
-ENTRYPOINT ["/app/audio_to_text_backend.py"]
+ENTRYPOINT ["/app/reel/backend_server.py"]
