+++ reel/grpc_server.py
@@ -20,7 +20,7 @@ from __future__ import annotations
 
 import sys
 
-from reel.grpc.server import main
+from reel.audio_grpc.server import main
 
 
 if __name__ == "__main__":
