+++ reel/tests/test_audio_to_text_grpc.py
@@ -19,9 +19,9 @@ import pytest
 from grpc_health.v1 import health_pb2
 from grpc_health.v1 import health_pb2_grpc
 
-from reel.grpc import audio_to_text_pb2
-from reel.grpc import audio_to_text_pb2_grpc
-import reel.grpc.server as audio_to_text_grpc_server
+from reel.audio_grpc import audio_to_text_pb2
+from reel.audio_grpc import audio_to_text_pb2_grpc
+import reel.audio_grpc.server as audio_to_text_grpc_server
 
 TEST_CERT_PEM = b"""-----BEGIN CERTIFICATE-----
 MIIDCTCCAfGgAwIBAgIUT5mIG80bR112155jkEP7GIhvy2cwDQYJKoZIhvcNAQEL
@@ -217,6 +217,13 @@ def start_server(
 ) -> subprocess.Popen[str]:
     """Start the gRPC server subprocess in test mode."""
     env = os.environ.copy()
+    # Set PYTHONPATH to parent of repo_root so 'reel' package can be found
+    parent_root = repo_root.parent
+    existing_pythonpath = env.get("PYTHONPATH", "")
+    if existing_pythonpath:
+        env["PYTHONPATH"] = f"{parent_root}{os.pathsep}{existing_pythonpath}"
+    else:
+        env["PYTHONPATH"] = str(parent_root)
     if test_mode:
         env["AUDIO_TO_TEXT_GRPC_TEST_MODE"] = "1"
     else:
@@ -226,7 +233,7 @@ def start_server(
     command = [
         sys.executable,
         "-m",
-        "audio_to_text_grpc.server",
+        "reel.audio_grpc.server",
         "--host",
         "127.0.0.1",
         "--port",
@@ -252,9 +259,16 @@ def run_grpc_command(
 ) -> subprocess.CompletedProcess[str]:
     """Run the gRPC server CLI and capture output."""
     env = os.environ.copy()
+    # Set PYTHONPATH to parent of repo_root so 'reel' package can be found
+    parent_root = repo_root.parent
+    existing_pythonpath = env.get("PYTHONPATH", "")
+    if existing_pythonpath:
+        env["PYTHONPATH"] = f"{parent_root}{os.pathsep}{existing_pythonpath}"
+    else:
+        env["PYTHONPATH"] = str(parent_root)
     if extra_env:
         env.update(extra_env)
-    command = [sys.executable, "-m", "audio_to_text_grpc.server", *args]
+    command = [sys.executable, "-m", "reel.audio_grpc.server", *args]
     return subprocess.run(
         command,
         cwd=repo_root,
@@ -1225,7 +1239,7 @@ def test_audio_to_text_grpc_rejects_tls_config(tmp_path: Path) -> None:
         [
             sys.executable,
             "-m",
-            "audio_to_text_grpc.server",
+            "reel.audio_grpc.server",
             "--tls-cert",
             str(cert_path),
         ],
